+++ src/App.vue
@@ -93,16 +93,6 @@ export default {
     if (this.url && this.token) {
       this.init()
     }
-    if (process.env.NODE_ENV === 'production' && process.env.GA_TRACKING_CODE) {
-      /* eslint-disable */
-      (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-      (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-      m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-      })(window,document,'script','https://www.google-analytics.com/analytics.js','ga');
-      ga('create', process.env.GA_TRACKING_CODE, 'auto');
-      ga('send', 'pageview');
-      /* eslint-enable */
-    }
   }
 }
 </script>